dashboard/configs: couple of fixes to baseline generation
Current script is not properly handling dependency chains longer
than two. Also Kconfig files that are read only if certain option
is enabled are not handled correctly. This patch is fixing these.

Also generation of baselines is not done using latest clang. This
is needed to ensure KCSAN and KMSAN are kept enabled.