pkg/kconfig: add simple config minimization
Minimize finds an equivalent with respect to the provided predicate, but smaller config.
It accepts base (small) and full (large) config. It is assumed that the predicate returns true for the full config.
It is also assumed that base and full are not just two completely arbitrary configs, but full it produced from base
mostly by adding more configs. The minimization procedure thus consists of figuring out what set of configs that
are present in full and are not present in base affect the predicate.

Update #2171