sys/freebsd: add freebsd12_closefrom and close_range
close_range is generally equivalent to the Linux syscall by the same
name, except that FreeBSD does not currently have any valid flags
defined for it.

Prior to FreeBSD 13, closefrom was redefined in terms of close_range in
libc and closefrom was pushed to compat12. closefrom(fd) was generally
equivalent to `close_range(fd, ~0U, 0)`.