dashboard/config/linux: update to next-20201124 with KFENCE enabled
Update configs to be based on next-20201124.

With that update, we can enable KFENCE to test KFENCE itself. Note, when
enabled on the KASAN config, we do not expect KFENCE reports to surface,
since KASAN's instrumentation will check an access before the actual
access (which may fault and be handled by KFENCE). Enabling KFENCE is
purely for the benefit of testing KFENCE.

Other notable changes coming to v5.11:
- WIRELESS_WDS subsystem removed
- DLCI subsystem removed
- X25_ASY subsystem removed
- NULL_TTY subsystem removed