pkg/report: skip drivers/usb/core/urb.c as guilty file
We are getting lots of WARNINGs in urb.c and all of them seem
to mean a bug in a particular driver. And fixes for these bugs
go into a particular driver code. But we send all of them to
urb.c maintainers. Skip urb.c as a guilty file.
If a bug happens to be in urb.c for real, a driver maintainers
should CC USB core maintainers on it.

Update #2284