pkg/build: add a new way of building linux images
Add a new, simpler procedure:
working bootable image is accepted as an input
and we only replace the kernel in it.
This just looks much saner and the right way to do it
(instead of assembling the image from bits and shelling out
to mkfs, fdisk, grub, etc). The new way also works inside of
containers much better.