sys/syz-sysgen: restore detection of unsupported kvm syscalls
The workaround for broken kvm kernel headers broke detection of unsupported syscalls.
The diagnostic check checks that a syscall is unsupported on all arches,
but arm/riscv kvm syscalls were not noted as unsupported (we dropped them before
giving them to the compiler). As the result it looked like kvm syscalls
are not unsupported on arm/riscv, and we did not produce the diagnostic.