pkg/osutil: use memfd_create on Linux
Mmapping plain files as shared memory leads to extra burden on the
system being fuzzed, because the OS will periodically sync that data to
the disk.

On Linux, use memfd_create to obtain a file handle for the shared memory
region. Experiments have demonstrated than it leads to 5-10% increase in
the fuzzing performance.