pkg/cover/backend: recover from DWARF parsing panic
GCC11 uses DWARF 5 format by default, which is not supported by go1.14 -
the Go version recommended by our documentation. This was fixed in
go1.16, but that version is not yet in the official stable Debian
packages.

Recover from such panic and return a meaningful error. This should help
the users to resolve the problem themselves.