tools/syz-kconf: better handle whitespaces for verbatim blocks
When we appended verbatim blocks we added '\n' for each of them,
(for every config file effectively). Even for empty verbatim blocks.
This worked fine because we had only 1 verbatim block and we do
TrimSpace for the resulting concatenated verbatim block
(so all excessive '\n' were removed). But this does not work for
multiple verbatim blocks because excessive '\n' will be in the middle
and won't be removed by TrimSpace.
Don't add '\n' for empty verbatim blocks.