tools/docker/old-env: don't use --allow-releaseinfo-change
This flag is needed for Bullseye that we use in env/syzbot containers.
I proactively added it to old-env as well assuming it's version-independent flag.

But Ubuntu 16.04 does not like this flag:

E: Command line option --allow-releaseinfo-change is not understood in combination with the other options
The command '/bin/sh -c apt-get update --allow-releaseinfo-change' returned a non-zero code: 100

Remove it from old-env.