sys/linux: neutralize sched_setattr
Setting itself or another process as a real-time one leads to the
starvation of kernel threads and, as a result, to false positive stall
bug reports. We have been getting complaints about them for already
quite a long time now.

Neutralize the policy argument of the syscall as much as possible given
the set of possible syzkaller mutations.